backend/requirements.txt
 fastapi
uvicorn[standard]
pydantic[email]
motor
python-dotenv
